requirements/base.txt
Django==4.0.2
Pillow==9.0.1
django-reversion==4.0.2
djangorestframework==3.13.1
PyJWT==2.3.0
django-ckeditor==6.2.0
drf-yasg==1.20.0
psycopg2==2.9.3
